sql/sql_analyse.cc:
  Auto merged
sql/sql_select.cc:
  Auto merged

+++ sql/handler.cc
@@ -883,13 +883,13 @@ int ha_create_table(const char *name, HA_CREATE_INFO *create_info,
 void ha_key_cache(void)
 {
   if (keybuff_size)
-    (void) init_key_cache(keybuff_size);
+    (void) init_key_cache((ulong) keybuff_size);
 }
 
 
 void ha_resize_key_cache(void)
 {
-  (void) resize_key_cache(keybuff_size);
+  (void) resize_key_cache((ulong) keybuff_size);
 }
